[NayNay] Transfer updates + unit testing
closes #124
