+++ src/flows/entropyTransfer/types.ts
@@ -1,0 +1,6 @@
+export interface TransferOptions { 
+  fromAddress?: string
+  fromPair?: any
+  to: string
+  amount: number | string | bigint 
+}
